+++ Code/GraphMol/MolDraw2D/MolDraw2DSVG.cpp
@@ -129,7 +129,7 @@ void MolDraw2DSVG::drawWavyLine(const Point2D &cds1, const Point2D &cds2,
   Point2D c1 = getDrawCoords(cds1);
 
   std::string col = DrawColourToSVG(colour());
-  unsigned int width = getDrawLineWidth();
+  double width = getDrawLineWidth();
   d_os << "<path ";
   if (d_activeClass != "") {
     d_os << "class='" << d_activeClass << "' ";
@@ -205,7 +205,7 @@ void MolDraw2DSVG::drawLine(const Point2D &cds1, const Point2D &cds2) {
   Point2D c1 = getDrawCoords(cds1);
   Point2D c2 = getDrawCoords(cds2);
   std::string col = DrawColourToSVG(colour());
-  unsigned int width = getDrawLineWidth();
+  double width = getDrawLineWidth();
   std::string dashString = "";
   const DashPattern &dashes = dash();
   if (dashes.size()) {
